Split MCP model surface tests
